% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_java.R
\name{check_java}
\alias{check_java}
\title{Check if Java is Installed and Correct Version}
\usage{
check_java(verbose = TRUE)
}
\arguments{
\item{verbose}{If \code{TRUE} messages are printed even when everything is
installed.}
}
\description{
Checks that Java is installed and of the correct version.
}
\examples{
\dontrun{
check_java()
}
}
\keyword{java}

