% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_setup.R
\name{check_setup}
\alias{check_setup}
\title{Check the Java/Stanford coreNLP Setup}
\usage{
check_setup()
}
\description{
Check to make sure that Java is installed and of correct version and that
Stanford's coreNLP is installed and in root.
}
\examples{
\dontrun{
check_setup()
}
}
\keyword{setup}

