% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreNLP_loc.R
\name{coreNLP_loc}
\alias{coreNLP_loc}
\alias{coreNLP_url}
\title{Make User Specific Path/URL to coreNLP}
\usage{
coreNLP_loc(version = stansent::coreNLP_version)

coreNLP_url(version = stansent::coreNLP_version)
}
\arguments{
\item{version}{The verson of coreNLP.}
}
\value{
returns a path as a string.
}
\description{
Make a user specific path/URL to coreNLP.
}
\examples{
coreNLP_loc()
}

