% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreNLP_version.R
\docType{data}
\name{coreNLP_version}
\alias{coreNLP_version}
\title{Stanford coreNLP Version}
\format{An object of class \code{character} of length 1.}
\usage{
coreNLP_version
}
\description{
A constant stating the version of coreNLP used as the package default.
}
\examples{
coreNLP_version
}
\references{
\url{http://stanfordnlp.github.io/CoreNLP/index.html#download}
}
\keyword{datasets}

