% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_version.R
\docType{data}
\name{java_version}
\alias{java_version}
\title{Java Version}
\format{An object of class \code{character} of length 1.}
\usage{
java_version
}
\description{
A constant stating the minimal version of Java needed.
}
\examples{
java_version
}
\keyword{datasets}

