% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_stanford_by.R
\name{plot.sentiment_stanford_by}
\alias{plot.sentiment_stanford_by}
\title{Plots a sentiment_stanford_by object}
\usage{
\method{plot}{sentiment_stanford_by}(x, max = 20, ...)
}
\arguments{
\item{x}{The sentiment_stanford_by object.}

\item{max}{The maximum point size.}

\item{\ldots}{ignored}
}
\value{
Returns a \pkg{ggplot2} object.
}
\description{
Plots a sentiment_stanford_by object.
}

