% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreNLPsetup_rexports.R,
%   R/sentimentr_reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{check_setup}
\alias{reexports}
\alias{uncombine}
\alias{reexports}
\alias{get_sentences}
\alias{reexports}
\alias{get_sentences2}
\alias{reexports}
\alias{highlight}
\alias{reexports}
\alias{average_downweighted_zero}
\alias{reexports}
\alias{average_weighted_mixed_sentiment}
\alias{reexports}
\alias{average_mean}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{coreNLPsetup}{\code{\link[coreNLPsetup]{check_setup}}}

  \item{sentimentr}{\code{\link[sentimentr]{uncombine}}, \code{\link[sentimentr]{get_sentences}}, \code{\link[sentimentr]{get_sentences2}}, \code{\link[sentimentr]{highlight}}, \code{\link[sentimentr]{average_downweighted_zero}}, \code{\link[sentimentr]{average_weighted_mixed_sentiment}}, \code{\link[sentimentr]{average_mean}}}
}}

