% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_stanford_by.R
\name{sentiment_stanford_by}
\alias{sentiment_stanford_by}
\title{Polarity Score (Sentiment Analysis) By Groups}
\usage{
sentiment_stanford_by(text.var, by = NULL,
  averaging.function = stansent::average_downweighted_zero, group.names, ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{by}{The grouping variable(s).  Default \code{NULL} uses the original
row/element indices; if you used a column of 12 rows for \code{text.var}
these 12 rows will be used as the grouping variable.  Also takes a single
grouping variable or a list of 1 or more grouping variables.}

\item{averaging.function}{A function for performing the group by averaging.
The default, \code{\link[sentimentr]{average_downweighted_zero}}, downweights
zero values in the averaging.  Note that the function must handle
\code{NA}s.  The \pkg{sentimentr} functions
\code{average_weighted_mixed_sentiment} and \code{average_mean} are also
available.  The former upweights negative when the analysts suspects the
speaker is likely to surround negatives with positives (mixed) as a polite
social convention but still the affective state is negative.  The later is a
standard mean average.}

\item{group.names}{A vector of names that corresponds to group.  Generally
for internal use.}

\item{\ldots}{Other arguments passed to \code{\link[stansent]{sentiment_stanford}}.}
}
\value{
Returns a \pkg{data.table} with grouping variables plus:
\itemize{
  \item  element_id - The id number of the original vector passed to \code{sentiment}
  \item  sentence_id - The id number of the sentences within each \code{element_id}
  \item  word_count - Word count \code{\link[base]{sum}}med by grouping variable
  \item  sd - Standard deviation (\code{\link[stats]{sd}}) of the sentiment/polarity score by grouping variable
  \item  ave_sentiment - Sentiment/polarity score \code{\link[base]{mean}} average by grouping variable
}
}
\description{
Approximate the sentiment (polarity) of text by grouping variable(s).
}
\details{
Note that the coreNLP is expensive with regard to time.  Typically it is
better to use \code{sentiment_stanford} and do the merging and aggregation/summarization
yourself.
}
\keyword{group}
\keyword{polarity,}
\keyword{sentiment,}
