% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansent-package.R
\docType{package}
\name{stansent}
\alias{stansent}
\alias{package-stansent}
\alias{stansent-package}
\title{Wrapper of Stanford's Sentiment Tagger}
\description{
Wraps Stanford's sentiment tagger in a way that makes the process easier to get set up.
}
\references{
http://nlp.stanford.edu/software/corenlp.shtml
}
