% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stanford.R
\name{check_stanford_installed}
\alias{check_stanford_installed}
\title{Check if Stanford coreNLP is Installed and In Root}
\usage{
check_stanford_installed(stanford = "stanford-corenlp-full-2015-04-20",
  download = "http://nlp.stanford.edu/software/stanford-corenlp-full-2015-04-20.zip")
}
\arguments{
\item{stanford}{The version of Stanford's coreNLP.}

\item{download}{The download url for Stanford's coreNLP.}
}
\description{
Checks that Stanford coreNLP is installed and in root.
}
\examples{
\dontrun{
check_stanford()
}
}
\references{
http://nlp.stanford.edu/software/corenlp.shtml
}
\keyword{stanford}

