% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr_reexports.R
\docType{import}
\name{reexports}
\alias{get_sentences}
\alias{get_sentences2}
\alias{highlight}
\alias{reexports}
\alias{uncombine}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sentimentr}{\code{\link[sentimentr]{uncombine}}, \code{\link[sentimentr]{get_sentences}}, \code{\link[sentimentr]{get_sentences2}}, \code{\link[sentimentr]{highlight}}}
}}
\keyword{internal}

